#5785 Extract RigFemClosestResultIndexCalculator class to separate file.
