+++ ApplicationCode/GeoMech/GeoMechDataModel/RigFemPartResultsCollection.cpp
@@ -3545,112 +3545,6 @@ int RigFemPartResultsCollection::partCount() const
 }
 
 //--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-RigFemClosestResultIndexCalculator::RigFemClosestResultIndexCalculator( RigFemPart*         femPart,
-                                                                        RigFemResultPosEnum resultPosition,
-                                                                        int                 elementIndex,
-                                                                        int                 m_face,
-                                                                        const cvf::Vec3d&   intersectionPointInDomain )
-{
-    m_resultIndexToClosestResult = -1;
-    m_closestNodeId              = -1;
-    m_closestElementNodeResIdx   = -1;
-
-    if ( resultPosition != RIG_ELEMENT_NODAL_FACE || m_face == -1 )
-    {
-        RigElementType elmType      = femPart->elementType( elementIndex );
-        const int*     elmentConn   = femPart->connectivities( elementIndex );
-        int            elmNodeCount = RigFemTypes::elmentNodeCount( elmType );
-
-        // Find the closest node
-        int   closestLocalNode = -1;
-        float minDist          = std::numeric_limits<float>::infinity();
-        for ( int lNodeIdx = 0; lNodeIdx < elmNodeCount; ++lNodeIdx )
-        {
-            int        nodeIdx         = elmentConn[lNodeIdx];
-            cvf::Vec3f nodePosInDomain = femPart->nodes().coordinates[nodeIdx];
-            float      dist            = ( nodePosInDomain - cvf::Vec3f( intersectionPointInDomain ) ).lengthSquared();
-            if ( dist < minDist )
-            {
-                closestLocalNode = lNodeIdx;
-                minDist          = dist;
-            }
-        }
-
-        if ( closestLocalNode >= 0 )
-        {
-            int nodeIdx = elmentConn[closestLocalNode];
-            m_closestElementNodeResIdx =
-                static_cast<int>( femPart->elementNodeResultIdx( elementIndex, closestLocalNode ) );
-
-            if ( resultPosition == RIG_NODAL )
-            {
-                m_resultIndexToClosestResult = nodeIdx;
-            }
-            else if ( resultPosition == RIG_ELEMENT_NODAL_FACE )
-            {
-                m_resultIndexToClosestResult = -1;
-            }
-            else if ( resultPosition == RIG_ELEMENT )
-            {
-                m_resultIndexToClosestResult = elementIndex;
-            }
-            else
-            {
-                m_resultIndexToClosestResult = m_closestElementNodeResIdx;
-            }
-
-            m_closestNodeId = femPart->nodes().nodeIds[nodeIdx];
-        }
-    }
-    else if ( m_face != -1 )
-    {
-        int elmNodFaceResIdx = -1;
-        int closestNodeIdx   = -1;
-        {
-            int closestLocFaceNode  = -1;
-            int closestLocalElmNode = -1;
-            {
-                RigElementType elmType        = femPart->elementType( elementIndex );
-                const int*     elmNodeIndices = femPart->connectivities( elementIndex );
-                int            faceNodeCount  = 0;
-                const int*     localElmNodeIndicesForFace =
-                    RigFemTypes::localElmNodeIndicesForFace( elmType, m_face, &faceNodeCount );
-
-                float minDist = std::numeric_limits<float>::infinity();
-                for ( int faceNodIdx = 0; faceNodIdx < faceNodeCount; ++faceNodIdx )
-                {
-                    int        nodeIdx         = elmNodeIndices[localElmNodeIndicesForFace[faceNodIdx]];
-                    cvf::Vec3f nodePosInDomain = femPart->nodes().coordinates[nodeIdx];
-                    float      dist = ( nodePosInDomain - cvf::Vec3f( intersectionPointInDomain ) ).lengthSquared();
-                    if ( dist < minDist )
-                    {
-                        closestLocFaceNode  = faceNodIdx;
-                        closestNodeIdx      = nodeIdx;
-                        closestLocalElmNode = localElmNodeIndicesForFace[faceNodIdx];
-                        minDist             = dist;
-                    }
-                }
-            }
-
-            int elmNodFaceResIdxElmStart  = elementIndex * 24; // HACK should get from part
-            int elmNodFaceResIdxFaceStart = elmNodFaceResIdxElmStart + 4 * m_face;
-
-            if ( closestLocFaceNode >= 0 )
-            {
-                elmNodFaceResIdx = elmNodFaceResIdxFaceStart + closestLocFaceNode;
-                m_closestElementNodeResIdx =
-                    static_cast<int>( femPart->elementNodeResultIdx( elementIndex, closestLocalElmNode ) );
-            }
-        }
-
-        m_resultIndexToClosestResult = elmNodFaceResIdx;
-        m_closestNodeId              = femPart->nodes().nodeIds[closestNodeIdx];
-    }
-}
-
-//--------------------------------------------------------------------------------------------------
 /// Internal functions
 //--------------------------------------------------------------------------------------------------
