+++ ApplicationCode/UserInterface/RiuFemTimeHistoryResultAccessor.cpp
@@ -19,6 +19,7 @@
 
 #include "RiuFemTimeHistoryResultAccessor.h"
 
+#include "RigFemClosestResultIndexCalculator.h"
 #include "RigFemPart.h"
 #include "RigFemPartCollection.h"
 #include "RigFemPartGrid.h"
